backend/requirements.txt
amqp==5.3.1
anthropic>=0.39.0
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
billiard==4.2.1
black==25.1.0
boto3==1.38.22
botocore==1.38.22
Brotli==1.1.0
celery==5.5.3
certifi==2025.4.26
cffi==1.17.1
channels>=4.0.0
channels-redis>=4.2.0
chardet==5.2.0
charset-normalizer==3.4.2
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==45.0.3
cssselect2==0.8.0
dj-rest-auth>=6.0.0
Django==5.1.7
django-allauth>=0.63.0
django-cors-headers==4.7.0
django-elasticsearch-dsl>=8.0.0
django-extensions==4.1
django-filter==25.1
django-hijack>=3.5.0
django-import-export>=4.0.0
django-model-utils==5.0.0
django-redis==5.4.0
django-rest-framework==0.1.0
django-storages==1.14.6
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
drf-extensions==0.8.0
elasticsearch-dsl>=8.0.0
fonttools==4.58.4
geographiclib==2.0
geopy==2.4.1
gunicorn==23.0.0
h11==0.14.0
html5lib==1.1
idna==3.10
jmespath==1.0.1
kombu==5.5.4
lxml==5.4.0
mypy-extensions==1.0.0
oscrypto==1.3.0
packaging==24.2
pathspec==0.12.1
pillow==11.1.0
platformdirs==4.3.7
prompt_toolkit==3.0.51
psycopg==3.2.6
psycopg-binary==3.2.6
psycopg2-binary==2.9.10
pycparser==2.22
pydyf==0.11.0
pyHanko==0.29.0
pyhanko-certvalidator==0.27.0
PyJWT==2.9.0
pypdf==5.6.0
pyphen==0.17.2
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.1
PyYAML==6.0.2
qrcode==8.2
redis==5.2.1
reportlab==4.4.1
requests==2.32.3
s3transfer==0.13.0
sentry-sdk>=2.0.0
six==1.17.0
sqlparse==0.5.3
svglib==1.5.1
tinycss2==1.4.0
tinyhtml5==2.0.0
typing_extensions==4.13.0
tzdata==2025.2
tzlocal==5.3.1
uritools==5.0.0
urllib3==2.4.0
uvicorn==0.34.0
vine==5.1.0
wcwidth==0.2.13
weasyprint==65.1
webencodings==0.5.1
whitenoise==6.9.0
xhtml2pdf==0.2.17
zopfli==0.2.3.post1
